requirements.txt
google-generativeai
pydantic
python-dotenv
pyyaml
